+++ src/global.d.ts
@@ -13,6 +13,7 @@ import type en_search from "../messages/en/search.json";
 import type en_settings from "../messages/en/settings.json";
 import type en_validation from "../messages/en/validation.json";
 import type en_admin from "../messages/en/admin.json";
+import type en_spec from "../messages/en/spec.json";
 
 type Messages = typeof en_common &
   typeof en_auth &
@@ -28,7 +29,8 @@ type Messages = typeof en_common &
   typeof en_search &
   typeof en_settings &
   typeof en_validation &
-  typeof en_admin;
+  typeof en_admin &
+  typeof en_spec;
 
 declare module "next-intl" {
   interface AppConfig {